Moved tailoredprofile test cases to new tailoring_tests suite
